fixes #848 server hang waiting for client handshake fixes #698 Need TCP stats fixes #699 Need IPC stats fixes #701 Need TLS stats
This commit addresses a problem when negotiating using one of the stream
based negotiation APIs -- a slow or misbehaving peer can prevent well
behaved ones from establishing a connection.  The fix is a fairly
significant change in how these transports link up, and it does rely
on the fact that the socket only has a single accept() or connect()
pending at a time (on a given endpoint that is).

While here, we have completely revamped the way transport statistics are
done, offering a standard API for collecting these statistics.

Unfortunately, this completely borks the statistics for inproc.  As we
are planning to change the way inproc works soon, in order to provide
more control and work on performance fixes for the message queue, we feel
this is an acceptable trade-off.  Furthermore, almost nobody uses inproc
for anything, and even fewer people are making use of the statistics
at this time.