fixes #1142 raw mode use of message headers is inconsistent
This correctly moves the entire protocol header for XREQ and XRESPONDENT
protocols to the message header (not the body).  This is where it should
always have been.  There is some small chance that applications which were
coded to parse the header from the body will break.  We don't think there
are any such applications in use.