fixes #1151 minor pairv1 optimizations possible
We've also finished fleshing out the statistics for pairv1, addressed a
possible race around the TTL, and added additional test cases to validate
things.  We also have tightened up the send logic to ensure we do not
send a bad message on the wire (drop the malformed message on the send side).